third_party/WebKit/LayoutTests/media/audio-constructor-src.html

<!DOCTYPE html>
<title>Test that Audio("url") constructor loads the specified resource.</title>
<script src="media-file.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
async_test(function(t) {
    var audioFile = findMediaFile("audio", "content/test");
    var audio = new Audio(audioFile);

    audio.onloadstart = t.step_func(function () {
        var url = audio.currentSrc;
        assert_equals(url.substr(url.lastIndexOf("/media/")+7), audioFile);
    });

    audio.oncanplaythrough = t.step_func_done();
});
</script>
